fix: LLVM backend — RuntimeCall &(T){v} args, socket fd widening
Builds on prior commit. Net LLVM PASS now ~767/814 (94.2%) from 710 baseline.

Changes:

1. **emit_global_runtime_init** — parse `&(TYPE){VAL}` compound-literal args
   alongside `sizeof(TYPE)` and numeric literals. Emits `alloca TYPE` +
   `store VAL` + passes the ptr. Fixes `Mutex[T](v)` / `Shared[T](v)` /
   similar concurrency constructors whose GIR expression is
   `gorget_mutex_new(sizeof(int64_t), &(int64_t){0})` — previously the
   LLVM parser fell back to `i64 8` (the literal `8` size) for both
   arguments, passing `8` as a pointer to the mutex's `memcpy` init.
   Unblocks thread_mutex, sync_condvar, shared_await_release.

2. **split_top_level_args** helper — splits at top-level commas only
   (ignoring nested parens/braces), so `gorget_mutex_new(sizeof(T), &(T){0})`
   tokenizes correctly. Previously `.split(',')` would split inside the
   compound literal.

3. **c_type_to_llvm** helper — small C→LLVM type-name mapper used by the
   compound-literal emission. Covers int8…int64, char, bool, float, double.

4. **Socket fd widening** — `GorgetSocket`, `GorgetServerSocket`,
   `GorgetUdpSocket`, `GorgetTlsSocket`, `GorgetTlsServerSocket` all had
   `int fd` (4 bytes) in the C runtime, but the Gorget source declares
   `struct Socket: int _handle` which maps to i64 (8 bytes). The LIR uses
   the Gorget semantics, so the LLVM-emitted `%Socket = { i64 }` had a
   different layout from the C-defined `GorgetSocket = { int }`. Result
   structs containing Socket had payloads at different offsets in the two
   backends, and runtime drop functions compiled against the C layout
   crashed when invoked by the LLVM-emitted code. Fix: widen `fd` to
   `int64_t` in the runtime so both backends see an 8-byte handle. Unblocks
   socket_connect and the socket-based tests.

Self-host pipeline aligned; integration tests 969/969 pass.

+++ src/backend/c/c_runtime.rs
@@ -7719,7 +7719,7 @@ pub const SOCKET_RUNTIME: &str = r#"
 #include <sys/time.h>
 
 typedef struct {
-    int fd;
+    int64_t fd;
 } GorgetSocket;
 
 // Connect to host:port, returning a socket fd or -1 on error
@@ -8019,7 +8019,7 @@ static void gorget_socket_set_blocking(GorgetSocket* sock) {
 // ServerSocket Gorget type (which is co-registered in std.net.socket) can
 // be typedef'd to GorgetServerSocket even when SERVER_SOCKET_RUNTIME is not emitted.
 typedef struct {
-    int fd;
+    int64_t fd;
 } GorgetServerSocket;
 
 /* Non-blocking accept: try accept once.
@@ -8144,7 +8144,7 @@ pub const UDP_SOCKET_RUNTIME: &str = r#"
 #include <poll.h>
 
 typedef struct {
-    int fd;
+    int64_t fd;
 } GorgetUdpSocket;
 
 typedef struct {
@@ -8337,7 +8337,7 @@ pub const TLS_SOCKET_RUNTIME: &str = r#"
 #include <errno.h>
 
 typedef struct {
-    int fd;
+    int64_t fd;
     SSL_CTX* ctx;
     SSL* ssl;
 } GorgetTlsSocket;
@@ -8532,7 +8532,7 @@ static void gorget_tls_set_timeout(GorgetTlsSocket* sock, int64_t ms) {
 // (e.g. via xtd.http) can reference TlsServerSocket in type aliases and Result
 // structs even when TLS_SERVER_RUNTIME is not emitted (no tls_server_bind call).
 typedef struct {
-    int fd;
+    int64_t fd;
     SSL_CTX* ctx;
 } GorgetTlsServerSocket;
